restored

+++ ft_str_cut.c
@@ -6,16 +6,16 @@
 /*   By: dskrypny <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/22 18:22:52 by dskrypny          #+#    #+#             */
-/*   Updated: 2017/12/09 14:24:06 by dskrypny         ###   ########.fr       */
+/*   Updated: 2018/06/02 15:25:44 by dskrypny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_str_cut(char *str, int size)
+char	*ft_str_cut(char *str, size_t size)
 {
 	char	*res;
-	int		i;
+	size_t		i;
 
 	i = 0;
 	if (!(res = (char *)malloc(sizeof(char) * size)))
